# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OmniGest",
        "description": "Documentação da API Omnigest",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://sistema-generico-api.vercel.app/api/api"
        }
    ],
    "paths": {
...
        "/servicoTipo": {
            "post": {
                "tags": [
                    "ServicoTipo"
                ],
                "summary": "Cria um novo tipo de serviço",
                "operationId": "createServicoTipo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "des_servico_tipo_stp",
                                    "vlr_servico_tipo_stp",
                                    "id_centro_custo_stp"
                                ],
                                "properties": {
                                    "des_servico_tipo_stp": {
                                        "type": "string",
                                        "example": "Serviço Premium"
                                    },
                                    "vlr_servico_tipo_stp": {
                                        "type": "integer",
                                        "example": 150
                                    },
                                    "id_centro_custo_stp": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de serviço criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServicoTipo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/servicoTipo/{id_servico_tipo}": {
            "get": {
                "tags": [
                    "ServicoTipo"
                ],
                "summary": "Obtém um tipo de serviço pelo ID",
                "operationId": "getServicoTipo",
                "parameters": [
                    {
                        "name": "id_servico_tipo",
                        "in": "path",
                        "description": "ID do tipo de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do tipo de serviço",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServicoTipo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Tipo de serviço não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "ServicoTipo"
                ],
                "summary": "Atualiza um tipo de serviço",
                "operationId": "updateServicoTipo",
                "parameters": [
                    {
                        "name": "id_servico_tipo",
                        "in": "path",
                        "description": "ID do tipo de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "des_servico_tipo_stp",
                                    "vlr_servico_tipo_stp"
                                ],
                                "properties": {
                                    "des_servico_tipo_stp": {
                                        "type": "string",
                                        "example": "Serviço Atualizado"
                                    },
                                    "vlr_servico_tipo_stp": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "id_centro_custo_stp": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de serviço atualizado"
                    },
                    "400": {
                        "description": "Erro na atualização"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ServicoTipo"
                ],
                "summary": "Inativa um tipo de serviço",
                "operationId": "deleteServicoTipo",
                "parameters": [
                    {
                        "name": "id_servico_tipo",
                        "in": "path",
                        "description": "ID do tipo de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de serviço inativado"
                    },
                    "400": {
                        "description": "Erro ao inativar"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ServicoTipo": {
                "properties": {
                    "id_servico_tipo_stp": {
                        "type": "integer"
                    },
                    "id_empresa_stp": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "des_servico_tipo_stp": {
                        "description": "Descrição do tipo de serviço",
                        "type": "string"
                    },
                    "vlr_servico_tipo_stp": {
                        "description": "Valor do tipo de serviço",
                        "type": "integer"
                    },
                    "is_ativo_stp": {
                        "description": "Indica se o tipo de serviço está ativo",
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id_centro_custo_stp": {
                        "description": "Chave estrangeira referenciando o centro de custo",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}